storeindex.py
from dotenv import load_dotenv
import os
from src.helper import load_pdf_file, filter_to_minimal_docs, text_split, download_hugging_face_embeddings
from pinecone import Pinecone
from pinecone import ServerlessSpec 
from langchain_pinecone import PineconeVectorStore

load_dotenv()


PINECONE_API_KEY=os.getenv("PINE_CONE_API_KEY")
HF_TOKEN= os.getenv("HF_TOKEN")
GROQ_API_KEY=os.getenv("GROQ_API_KEY")

os.environ["HF_TOKEN"]=HF_TOKEN
os.environ["PINECONE_API_KEY"]=PINECONE_API_KEY
os.environ["GROQ_API_KEY"]=GROQ_API_KEY


extracted_data=load_pdf_files(data="data/")
minimal_docs= filter_to_minimal_docs(extracted_data)
texts_chunk= text_split(minimal_docs)

embeddings=download_embeddings()

pinecone_apikey=PINECONE_API_KEY
pc=Pinecone(api_key=pinecone_apikey)

index_name="medical-chatbot"

if not pc.has_index(index_name):
    pc.create_index(
        name=index_name,
        dimension=384,
        metric="cosine",
        spec=ServerlessSpec(cloud='aws',region='us-east-1')
    
    )
index=pc.Index(index_name)

docsearch = PineconeVectorStore.from_documents (
documents=texts_chunk, embedding=embeddings, index_name=index_name
)